image_classification/requirements.txt
packaging==24.1
numpy==1.21.6
pandas==1.3.5
tqdm==4.66.2
torch==1.13.1
torchvision==0.14.1
scikit_image==0.18.3
scikit_learn==1.0.2
reedsolo==1.7.0
Pillow==10.4.0
einops==0.8.0
matplotlib==3.5.3
